remove extra blank line in outline
You might feel that 50 characters is too short though, so another standard that you could use is to keep the summary to a max of 72, 78, or 80 chars.

For the rest of the commit message, keeping the max line length to 72, 78, or 80 chars max can also be helpful, like if you often split your monitor screens with a terminal in one half and a browser or editor in the other half. Many editors have shortcuts that will auto-wrap long lines to a max column length for you.